voice-agent-day2/requirements.txt
Flask
requests
python-dotenv
fastapi
uvicorn
google-generativeai

